Moved Enums into Widgets folder
